move modutils docs
