+++ specparam/tests/modutils/test_docs.py
@@ -1,37 +1,14 @@
-"""Tests for specparam.core.modutils.
+"""Tests for specparam.modutils.docs.
 
 Note: the decorators for copying documentation are not currently tested.
 """
 
 from pytest import raises
 
-from specparam.core.modutils import *
+from specparam.modutils.docs import *
 
 ###################################################################################################
 ###################################################################################################
-
-def test_safe_import():
-
-    np = safe_import('numpy')
-    assert np
-
-    bad = safe_import('bad')
-    assert not bad
-
-def test_check_dependency():
-
-    import numpy as np
-    @check_dependency(np, 'numpy')
-    def subfunc_good():
-        pass
-    subfunc_good()
-
-    bad = None
-    @check_dependency(bad, 'bad')
-    def subfunc_bad():
-        pass
-    with raises(ImportError):
-        subfunc_bad()
 
 def test_docs_drop_param(tdocstring):
